+++ src/components/loading/loading-spinner/loading-spinner.tsx
@@ -1,13 +1,10 @@
-import { Spinner } from "@shopify/polaris";
+import { Box, Spinner, VerticalStack } from "@shopify/polaris";
 import { memo } from "react";
 
 export const LoadingSpinner = memo(() => (
-  <div
-    style={{
-      display: "flex",
-      justifyContent: "center",
-    }}
-  >
-    <Spinner accessibilityLabel="Loading" hasFocusableParent={false} />
-  </div>
+  <Box padding="4">
+    <VerticalStack align="center" inlineAlign="center">
+      <Spinner accessibilityLabel="Loading" hasFocusableParent={false} />
+    </VerticalStack>
+  </Box>
 ));
