+++ src/pages/admin/users/shifts/delete-shift-group/lazy.component.tsx
@@ -11,22 +11,24 @@ export default () => {
   const { show } = useToast();
   const submit = useSubmit();
   const { formatInTimezone } = useDate();
+  const { isOpen, close, redirect } = useShiftModal();
   const { t } = useTranslation({
-    id: "delete-shift",
+    id: "delete-shift-group",
     locales,
   });
 
-  const { isOpen, close, redirect } = useShiftModal();
   const loaderData = useLoaderData() as Awaited<ReturnType<typeof loader>>;
 
   const onCancel = useCallback(() => {
-    redirect(`./../edit-shift`);
+    redirect(`./../edit/edit-shift-group`);
   }, [redirect]);
 
   const onDestroy = useCallback(async () => {
     submit(
       {
         userId: loaderData.userId,
+        start: loaderData.start.toJSON(),
+        end: loaderData.end.toJSON(),
       },
       {
         method: "delete",
@@ -34,7 +36,7 @@ export default () => {
     );
     close();
     show({ content: t("success") });
-  }, [loaderData.userId, close, show, submit, t]);
+  }, [close, loaderData, show, submit, t]);
 
   return (
     <Modal
@@ -56,8 +58,8 @@ export default () => {
       <Modal.Section>
         <Text as="p">
           {t("date", {
-            date: <strong>{formatInTimezone(loaderData?.start, "PPP")}</strong>,
-            day: <strong>{formatInTimezone(loaderData?.start, "EEEE")}</strong>,
+            from: <strong>{formatInTimezone(loaderData?.start, "PPP")}</strong>,
+            to: <strong>{formatInTimezone(loaderData?.end, "PPP")}</strong>,
           })}
           {", "}
           {t("time", {
@@ -72,17 +74,17 @@ export default () => {
 
 const locales = {
   da: {
-    date: "Du er i gang med at slette arbejdsdagen {day} d. {date}",
+    date: "Du er i gang med at slette arbejdsperioden fra d. {from} tilmed d. {to}",
     time: "fra kl. {from} til {to}",
-    action: "Slet arbejdsdag",
+    action: "Slet arbejdsperiode",
     title: "Er du sikker?",
     cancel: "Annullere",
     success: "Vagtplan slettet",
   },
   en: {
-    date: "Shiftday {day} og date {date}",
+    date: "You are deleting shifts from {from} to {to}",
     time: "Time from {from} til {to}",
-    action: "Delete shiftday",
+    action: "Delete shiftdays",
     title: "Are you sure?",
     cancel: "Cancel",
     success: "Shift deleted",
