+++ src/pages/admin/users/shifts/edit-shift-group/action.ts
@@ -2,9 +2,7 @@ import { AxiosError } from "axios";
 import { ActionFunctionArgs } from "react-router-dom";
 import {
   getUserShiftGetAllQueryKey,
-  getUserShiftGetByIdQueryKey,
   getUserShiftGetGroupQueryKey,
-  userShiftUpdate,
   userShiftUpdateGroup,
 } from "~/api/bookingShopifyApi";
 import { BadResponseResponse } from "~/api/model";
@@ -13,33 +11,20 @@ import { queryClient } from "~/providers/query-provider";
 
 export const action = async ({ request, params }: ActionFunctionArgs) => {
   const userId = params.userId || "";
-  const { selectedGroupId, selectedShiftId } = scheduleGetQueries(request.url);
+  const { selectedGroupId } = scheduleGetQueries(request.url);
 
   try {
     const formData = Object.fromEntries(await request.formData());
 
-    let response;
-    if (selectedGroupId) {
-      // form.days is coming as "wednesday,thursday", useSubmit from react-router convert that.
-      response = await userShiftUpdateGroup(userId, selectedGroupId, {
-        ...formData,
-        days: (formData.days as string).split(","),
-      } as never);
+    // form.days is coming as "wednesday,thursday", useSubmit from react-router convert that.
+    const response = await userShiftUpdateGroup(userId, selectedGroupId, {
+      ...formData,
+      days: (formData.days as string).split(","),
+    } as never);
 
-      queryClient.invalidateQueries(
-        getUserShiftGetGroupQueryKey(userId, selectedGroupId)
-      );
-    } else {
-      response = await userShiftUpdate(
-        userId,
-        selectedShiftId,
-        formData as never
-      );
-
-      queryClient.invalidateQueries(
-        getUserShiftGetByIdQueryKey(userId, selectedShiftId)
-      );
-    }
+    queryClient.invalidateQueries(
+      getUserShiftGetGroupQueryKey(userId, selectedGroupId)
+    );
 
     await queryClient.invalidateQueries(
       getUserShiftGetAllQueryKey(userId, null as never)
