+++ src/pages/admin/users/shifts/edit/index.ts
@@ -1,0 +1,1 @@
+export * from "./component";
