+++ src/pages/admin/users/show-user/component.tsx
@@ -1,79 +1,42 @@
-import { AlphaCard, Page } from "@shopify/polaris";
-import { Suspense } from "react";
+import { Page } from "@shopify/polaris";
 
 import { subject } from "@casl/ability";
-import { Outlet } from "react-router-dom";
+import { Outlet, useLoaderData } from "react-router-dom";
 import { BadgeStatus } from "~/components/badge-status";
-import { LoadingSpinner } from "~/components/loading/loading-spinner";
-import { ScheduleCalendar } from "~/components/schedule-calendar";
-import { Await, useDeferredLoaderData } from "~/lib/loader-data";
 import { useAbility } from "~/providers/ability-provider";
 import { useTranslation } from "~/providers/translate-provider";
 import { loader } from "./loader";
 
 export function Component() {
-  const { user, shifts } = useDeferredLoaderData<typeof loader>();
+  const user = useLoaderData() as Awaited<ReturnType<typeof loader>>;
   const ability = useAbility();
   const { t } = useTranslation({ id: "user-schedule", locales });
 
   return (
-    <>
-      <Suspense fallback={<LoadingSpinner />}>
-        <Await resolve={user}>
-          {(user) => (
-            <Page
-              fullWidth
-              title={t("title", user)}
-              titleMetadata={<BadgeStatus active={user?.active} />}
-              backAction={{ content: "user", url: "../users" }}
-              primaryAction={
-                ability.can("update", subject("user", user)) && {
-                  content: t("edit", user),
-                  url: `edit`,
-                }
-              }
-              secondaryActions={
-                ability.can("update", subject("user", user)) && [
-                  {
-                    content: t("add"),
-                    onAction: () => {},
-                  },
-                ]
-              }
-            >
-              <AlphaCard>
-                <Suspense fallback={<LoadingSpinner />}>
-                  <Await resolve={shifts}>
-                    {(shifts) => <ScheduleCalendar data={shifts} />}
-                  </Await>
-                </Suspense>
-              </AlphaCard>
-            </Page>
-          )}
-        </Await>
-      </Suspense>
+    <Page
+      fullWidth
+      title={t("title", user)}
+      titleMetadata={<BadgeStatus active={user?.active} />}
+      backAction={{ content: "user", url: "../users" }}
+      primaryAction={
+        ability.can("update", subject("user", user)) && {
+          content: t("edit"),
+          url: `edit`,
+        }
+      }
+    >
       <Outlet />
-    </>
+    </Page>
   );
 }
 
 const locales = {
   da: {
-    add: "Tilføj vagt",
-    edit: "Redigere {fullname}",
-    loading: {
-      data: "Henter medarbejder vagtplan",
-      user: "Henter medarbejder data",
-    },
+    edit: "Redigere bruger",
     title: "{fullname} vagtplan",
   },
   en: {
-    add: "Add shift",
-    edit: "Edit {fullname}",
-    loading: {
-      data: "Loading user shifts",
-      user: "Loading user data",
-    },
+    edit: "Edit user",
     title: "{fullname} shifts",
   },
 };
