+++ src/routes.tsx
@@ -29,7 +29,7 @@ const router = createBrowserRouter(
           lazy={() => import("./pages/admin/users/create-user")}
         />
         <Route
-          path="user/:userId/edit"
+          path="user/:userId/edit-user"
           lazy={() => import("./pages/admin/users/edit-user")}
         />
         <Route
@@ -37,31 +37,51 @@ const router = createBrowserRouter(
           lazy={() => import("./pages/admin/users/show-user")}
         >
           <Route
-            path="create"
-            lazy={() => import("./pages/admin/users/create")}
+            path="shifts"
+            lazy={() => import("./pages/admin/users/shifts")}
           >
             <Route
-              path="create-shift"
-              lazy={() => import("./pages/admin/users/create-shift")}
+              path="create"
+              lazy={() => import("./pages/admin/users/shifts/create")}
+            >
+              <Route
+                path="create-shift"
+                lazy={() => import("./pages/admin/users/shifts/create-shift")}
+              />
+              <Route
+                path="create-shift-group"
+                lazy={() =>
+                  import("./pages/admin/users/shifts/create-shift-group")
+                }
+              />
+            </Route>
+
+            <Route
+              path="edit"
+              lazy={() => import("./pages/admin/users/shifts/edit")}
+            >
+              <Route
+                path="edit-shift"
+                lazy={() => import("./pages/admin/users/shifts/edit-shift")}
+              />
+              <Route
+                path="edit-shift-group"
+                lazy={() =>
+                  import("./pages/admin/users/shifts/edit-shift-group")
+                }
+              />
+            </Route>
+            <Route
+              path="delete-shift"
+              lazy={() => import("./pages/admin/users/shifts/delete-shift")}
             />
             <Route
-              path="create-shift-group"
-              lazy={() => import("./pages/admin/users/create-shift-group")}
+              path="delete-shift-group"
+              lazy={() =>
+                import("./pages/admin/users/shifts/delete-shift-group")
+              }
             />
           </Route>
-
-          <Route
-            path="edit-shift"
-            lazy={() => import("./pages/admin/users/edit-shift")}
-          />
-          <Route
-            path="delete-shift"
-            lazy={() => import("./pages/admin/users/delete-shift")}
-          />
-          <Route
-            path="delete-shift-group"
-            lazy={() => import("./pages/admin/users/delete-shift-group")}
-          />
         </Route>
         <Route path="users" lazy={() => import("./pages/admin/users/list")} />
       </Route>
