+++ staticwebapp.config.json
@@ -1,0 +1,10 @@
+{
+  "routes": [
+    {
+      "route": "/*",
+      "serve": "/index.html",
+      "statusCode": 200
+    }
+  ],
+  "rewrites": []
+}
